Move FormWithQuickAdd to own directory
